requirements.txt
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.3
docopt==0.6.2
idna==2.8
numpy==1.21.1
pandas==1.2.3
pipreqs==0.4.10
PySimpleGUI==4.45.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.22.0
six==1.16.0
stdiomask==0.0.6
tabulate==0.8.9
urllib3==1.25.8
yarg==0.1.9
